Added cancel option and progress monitoring to copy and move, moved file opration task to different class

+++ src/main/java/pl/poznan/put/fc/tpal/jcommander/fileOperation/MoveFile.java
@@ -9,17 +9,18 @@ import java.nio.file.attribute.BasicFileAttributes;
 import java.util.List;
 
 import static java.nio.file.FileVisitResult.CONTINUE;
+import static java.nio.file.FileVisitResult.TERMINATE;
 
 /**
  * @author Kamil Walkowiak
  */
-public class MoveFiles extends FilesOperation {
+public class MoveFile extends FileOperation {
     private List<Path> sourcePaths;
     private Path targetPath;
     private Path currentSourcePath;
     private Path currentTargetPath;
 
-    public MoveFiles(List<File> files, BooleanProperty isCanceledProperty, List<Path> sourcePaths, Path targetPath) {
+    public MoveFile(List<File> files, BooleanProperty isCanceledProperty, List<Path> sourcePaths, Path targetPath) {
         super(files, isCanceledProperty);
         this.sourcePaths = sourcePaths;
         this.targetPath = targetPath;
@@ -28,6 +29,9 @@ public class MoveFiles extends FilesOperation {
     @Override
     public void execute() throws IOException {
         for(Path path: sourcePaths) {
+            if(isCanceledProperty.get()) {
+                break;
+            }
             currentSourcePath = path;
             currentTargetPath = Paths.get(targetPath.toString(), currentSourcePath.getFileName().toString());
             Files.walkFileTree(path, this);
@@ -36,25 +40,37 @@ public class MoveFiles extends FilesOperation {
 
     @Override
     public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
-        Path targetDir = currentTargetPath.resolve(currentSourcePath.relativize(dir));
-        try {
-            Files.copy(dir, targetDir);
-        } catch (FileAlreadyExistsException e) {
-            if (!Files.isDirectory(targetDir))
-                throw e;
+        if(!isCanceledProperty.get()) {
+            Path targetDir = currentTargetPath.resolve(currentSourcePath.relativize(dir));
+            try {
+                Files.copy(dir, targetDir);
+            } catch(FileAlreadyExistsException e) {
+                if(!Files.isDirectory(targetDir))
+                    throw e;
+            }
+            return CONTINUE;
+        } else {
+            return TERMINATE;
         }
-        return CONTINUE;
     }
 
     @Override
     public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-        Files.move(file, currentTargetPath.resolve(currentSourcePath.relativize(file)));
-        return CONTINUE;
+        if(!isCanceledProperty.get()) {
+            Files.move(file, currentTargetPath.resolve(currentSourcePath.relativize(file)));
+            return CONTINUE;
+        } else {
+            return TERMINATE;
+        }
     }
 
     @Override
     public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-        Files.delete(dir);
-        return CONTINUE;
+        if(!isCanceledProperty.get()) {
+            Files.delete(dir);
+            return CONTINUE;
+        } else {
+            return TERMINATE;
+        }
     }
 }
